backend_v2/backend/requirements/development.txt
-r base.txt

# Development and testing tools – not for production
django-debug-toolbar>=4.2,<5.0          # SQL panel, cache, etc.
pytest>=7.4,<8.0                        # Testing framework
pytest-django>=4.7,<5.0                 # Django plugin for pytest
factory-boy>=3.3,<4.0                    # Model factories for tests
coverage>=7.3,<8.0                       # Code coverage measurement
pytest-cov>=5.0,<6.0                      # Pytest integration for coverage
drf-spectacular>=0.27,<0.28